fix: Harden provider tool rendering and control requests
## Motivation

- A review of the previous commit found defects across every agent provider,
  the shared control-request path, and the storage schema. That commit also
  stopped short of its own goal: ZCode and Pi still rendered tool rows
  through their own local tables.

## Control requests

- Record every published control request in one registrar, and retire both
  the provider record and the browser card through one withdrawal. Five
  publishers had four partial retirements between them, so a stop answered
  the agent for a permission prompt and left an MCP elicitation blocked for
  the rest of the session.
- Canonicalize a JSON-RPC id once, so 12, 12.0 and 1.2e1 name one request.
  A cancel that spelled the id differently matched nothing.
- Answer an inbound request the worker does not support with -32601, rather
  than persisting it as a transcript row while the runtime waits.
- Restore the degraded plan-execution path. A ZCode or native Copilot plan
  approval that cleared context failed after the permission mode had already
  moved, and every retry repeated it.
- Keep the composer draft, and re-enable the decision buttons, when the
  worker records an answer without completing it.
- Clear the interrupt note when a session swap discards the turn that
  carried it, so the next turn stops marking finished tools interrupted.

## Providers

- Reset the whole session state when a Copilot goal clear replaces the
  session, and stop the connection when the reopen fails. The tab otherwise
  stayed busy forever, or addressed a session the runtime had destroyed.
- Start the Copilot reader only after the connection is adopted, which
  closes a nil dereference and a data race.
- Coalesce the Copilot background reads behind one keyed runner, so a burst
  of change events no longer races two settings snapshots.
- Treat a Pi session id and its file as one identity, so a replacement
  session stops persisting the previous file as its resume handle.
- Make the ZCode stopped-turn timer one critical section, so a cancel can no
  longer be undone into a spurious interruption 30 seconds later.
- Read every provider store through the StoredSessionQuery home seam.

## Tool rendering

- Move ZCode and Pi onto the shared ToolPresentation path. All nine
  providers now render tool rows through one component and derive their
  toolbar metadata from one place; four Pi tables and two ZCode switches are
  gone.
- Narrow ToolPresentation.kind to a closed union and make the icon, label
  and title switches exhaustive. The three tables disagreed about Reasonix's
  move kind, which drew a generic wrench and dumped raw JSON.
- Derive a retained row's outcome once, and draw the outcome header from one
  component that cannot be forgotten.
- Unwrap an Agent Client Protocol result envelope in one module. A wrapped
  response rendered as a raw JSON bubble where the turn-end divider belongs.

## Storage and contracts

- Give every enum column a CHECK that starts at 1, and let sqlc carry the
  proto enum types, so a swapped Kind and Status no longer compile.
- Repeat each partial index's predicate in every query that rides it, and
  pin all seven with one test.
- Move the Cursor methods, the Copilot session-event method, the Reasonix
  tool record and the ACP result role into contracts.
- Derive the enum-migration script's rebuild set from the schema diff, and
  bind its values as parameters. It could not convert a worker database.

## Efficiency

- Enrich tool supplements on a worker goroutine, so a slow provider database
  no longer back-pressures the agent's stdout pipe, and give the final pass
  its own longer budget.
- Return early from the message-metadata merge when a message carries no
  metadata: 13 allocations for each message become none.
- Keep one database handle and one parsed event graph for each session,
  instead of reopening and replaying from byte 0 on every message.

+++ backend/internal/worker/service/control_plan_session_test.go
@@ -117,7 +117,7 @@ func TestNativePlanRecordingDoesNotChangeAReplacementSession(t *testing.T) {
 	response := []byte(`{"response":{"request_id":"plan","response":{"behavior":"allow"}}}`)
 	claimed, err := svc.Output.claimControlResponseAnswer(db.ClaimControlResponseAnswerParams{
 		AgentID: "agent-1", RequestID: "plan", ClaimToken: "claim", AgentSessionID: "original",
-		AgentProvider: int64(old.AgentProvider), RequestPayload: []byte(`{"request":{"tool_name":"ExitPlanMode"}}`),
+		AgentProvider: old.AgentProvider, RequestPayload: []byte(`{"request":{"tool_name":"ExitPlanMode"}}`),
 		ResponseContent: response, ResolvedContent: response,
 		PlanApprovalSettings: []byte(`{"permissionMode":"bypassPermissions","clearContext":false}`),
 	})
